fix: resolve static analysis warnings for Addons validator
Fix uninitialized variables, incorrect PHPDoc tags, missing property
declarations, and type cast issues flagged by the Addons validator.
No behavioral changes.

+++ controllers/admin/AdminConfigureCiklikController.php
@@ -659,7 +659,7 @@ class AdminConfigureCiklikController extends ModuleAdminController
                 if ($shopData instanceof ShopData) {
                     Configuration::updateGlobalValue(Ciklik::CONFIG_HOST, $shopData->host);
 
-                    $webservice = new WebserviceKey(Configuration::get(Ciklik::CONFIG_WEBSERVICE_ID));
+                    $webservice = new WebserviceKey((int) Configuration::get(Ciklik::CONFIG_WEBSERVICE_ID));
 
                     $ciklikShopApi->metadata(
                         [
